requirements.txt
fastapi>=0.115.6
uvicorn[standard]>=0.34.0
python-dotenv>=1.0.1
pydantic>=2.10.4
langchain>=0.3.25
langchain-core>=1.2.5
langchain-community>=0.3.25
langchain-google-genai>=4.2.1
langchain-text-splitters>=0.3.8
supabase>=2.13.0
pypdf>=5.3.0
